requirements-test.txt
alabaster==0.7.7
Babel==2.2.0
coverage==4.0.3
decorator==4.0.9
docutils==0.12
Flask==0.10.1
httpbin==0.4.1
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
py==1.4.31
Pygments==2.1.1
PySocks==1.5.6
pytest==2.8.7
pytest-cov==2.2.1
pytest-httpbin==0.2.0
pytest-mock==0.11.0
pytz==2015.7
six==1.10.0
snowballstemmer==1.2.1
Sphinx==1.3.5
sphinx-rtd-theme==0.1.9
Werkzeug==0.11.4
wheel==0.29.0
mock
requests-mock
